plugin update

+++ org.kevoree.modeling.java2typescript.mavenplugin/src/main/java/org/kevoree/modeling/java2typescript/mavenplugin/TSCRunner.java
@@ -2,19 +2,22 @@ package org.kevoree.modeling.java2typescript.mavenplugin;
 
 import de.flapdoodle.embed.nodejs.*;
 import de.flapdoodle.embed.process.config.IRuntimeConfig;
+import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugin.logging.Log;
 
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
+import java.util.Arrays;
 
 /**
  * Created by duke on 08/12/14.
  */
 public class TSCRunner {
 
-    public static void run(File src, File target) throws IOException {
+    public static void run(File src, File target, File[] libraries, boolean copyLibDTs) throws Exception {
         IRuntimeConfig runtimeConfig = (new NodejsRuntimeConfigBuilder()).defaults().build();
         NodejsProcess node = null;
         ArrayList<String> paramsCol = new ArrayList<String>();
@@ -24,15 +27,27 @@ public class TSCRunner {
                 paramsCol.add(files[i].getAbsolutePath());
             }
         }
+        if(libraries != null) {
+            for (int i = 0; i < libraries.length; i++) {
+                File[] lib = libraries[i].listFiles();
+                for (int j = 0; j < lib.length; j++) {
+                    if (lib[j].getName().endsWith(".ts")) {
+                        paramsCol.add(lib[j].getAbsolutePath());
+                    }
+                }
+            }
+        }
 
         File targetTSCBIN = new File(target, "tsc.js");
         targetTSCBIN.deleteOnExit();
         Files.copy(TSCRunner.class.getClassLoader().getResourceAsStream("tsc.js"), targetTSCBIN.toPath(), StandardCopyOption.REPLACE_EXISTING);
 
-        File targetLIBD = new File(target, "lib.d.ts");
-        targetLIBD.deleteOnExit();
-        Files.copy(TSCRunner.class.getClassLoader().getResourceAsStream("lib.d.ts"), targetLIBD.toPath(), StandardCopyOption.REPLACE_EXISTING);
-        paramsCol.add(targetLIBD.getAbsolutePath());
+        if(copyLibDTs) {
+            File targetLIBD = new File(target, "lib.d.ts");
+            targetLIBD.deleteOnExit();
+            Files.copy(TSCRunner.class.getClassLoader().getResourceAsStream("lib.d.ts"), targetLIBD.toPath(), StandardCopyOption.REPLACE_EXISTING);
+            paramsCol.add(targetLIBD.getAbsolutePath());
+        }
 
         paramsCol.add("--outDir");
         paramsCol.add(target.getAbsolutePath());
@@ -40,12 +55,17 @@ public class TSCRunner {
         paramsCol.add("-d");
 
 
+        System.out.println("Command line:" + String.join("; ", paramsCol));
+
         NodejsConfig nodejsConfig = new NodejsConfig(NodejsVersion.Main.V0_10, targetTSCBIN.getAbsolutePath(), paramsCol, target.getAbsolutePath());
         NodejsStarter runtime = new NodejsStarter(runtimeConfig);
         try {
             NodejsExecutable e = (NodejsExecutable) runtime.prepare(nodejsConfig);
             node = (NodejsProcess) e.start();
-            node.waitFor();
+            int retVal = node.waitFor();
+            if(retVal != 0) {
+                throw new Exception("There were TypeScript compilation errors.");
+            }
         } catch (InterruptedException var11) {
             var11.printStackTrace();
         } finally {
